% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combineSlicedData}
\alias{combineSlicedData}
\title{Combine two \code{SlicedData} objects}
\usage{
combineSlicedData(...)
}
\arguments{
\item{...}{\code{SlicedData} objects}
}
\value{
A \code{SlicedData} object containg the data from \code{x} and \code{y}
}
\description{
Combine two \code{SlicedData} objects
}
\author{
Peter Humburg
}

