\name{combineSlicedData}
\alias{combineSlicedData}
\title{Combine two \code{SlicedData} objects}
\usage{
  combineSlicedData(x, y)
}
\arguments{
  \item{x}{A \code{SlicedData} object}

  \item{y}{A \code{SlicedData} object}
}
\value{
  A \code{SlicedData} object containg the data from
  \code{x} and \code{y}
}
\description{
  Combine two \code{SlicedData} objects
}
\author{
  Peter Humburg
}

