\name{toCubeX}
\alias{toCubeX}
\title{Extract named SNPs from SlicedData object and return them in the format
expected by CubeX}
\usage{
  toCubeX(data, snps)
}
\description{
  Extract named SNPs from SlicedData object and return them
  in the format expected by CubeX
}
\author{
  Peter Humburg
}
\keyword{internal}

